updated data to read floats

+++ activations.cpp
@@ -4,11 +4,11 @@
 #define relu(a) (a)>0?(a):0
 using namespace::std;
 
-double* activations( bool flag,  int numrows, int numcols, double *matrix){
+float* activations( bool flag,  int numrows, int numcols, float *matrix){
 	int m=numrows; //number of rows
 	int n=numcols; //number of columns
-  
-	double *output = new double[m*n];
+
+	float *output = new float[m*n];
 
 
 	if(flag==true){// if flag = true then compute tanh else relu
@@ -34,16 +34,16 @@ double* activations( bool flag,  int numrows, int numcols, double *matrix){
 
 /*
 int main(){
-  double** a = new double*[4];
+  float** a = new float*[4];
   for(int i = 0;i<4;i++){
-    a[i]= new double[4];
+    a[i]= new float[4];
   }
   for(int i = 0;i<4;i++){
     for(int j = 0;j<4;j++ ){
       a[i][j]=-(i+j)*1.0;
     }
   }
-  double **b ;
+  float **b ;
    activations(false,make_pair(4,4),a);
   cout<<a[0][0]<<" "<<a[0][1]<<endl<<a[1][0]<<" "<<a[1][1]<<endl;
   return 0;
