+++ activations.h
@@ -1,4 +1,4 @@
 #pragma once
 using namespace std;
-//double** activations(bool flag,  pair<const int, const int> rc_matrix, double **matrix);
-double* activations( bool flag,  int numrows, int numcols, double *matrix);
+//float** activations(bool flag,  pair<const int, const int> rc_matrix, float **matrix);
+float* activations( bool flag,  int numrows, int numcols, float *matrix);
