+++ convolute_multiple.h
@@ -1,4 +1,4 @@
 #pragma once
 using namespace:: std;
 
-double ** convolute_multiple(double** input_layers, int input_width, int input_channels,double*** filter_layers, double* biases, int filter_width,int output_channels,int pad_size,bool toPad, bool relu);
+float ** convolute_multiple(float** input_layers, int input_width, int input_channels,float*** filter_layers, float* biases, int filter_width,int output_channels,int pad_size,bool toPad, bool relu);
