+++ main.cpp
@@ -19,7 +19,7 @@
 #include "convolute_multiple.h"
 using namespace std;
 
-double* read_matrix(int *size, char *name){
+float* read_matrix(int *size, char *name){
 /*Reads (square_matrix) data given in the file with above name.*/
 /*modifies the size variable such that *size will store the row/column size*/
 
@@ -32,8 +32,8 @@ double* read_matrix(int *size, char *name){
     return NULL;
   }
 
-  vector <double> input_vec;
-  double val_read;
+  vector <float> input_vec;
+  float val_read;
   while(file>>val_read){
     input_vec.push_back(val_read);//"name" file should contain the matrix in row major order
   }
@@ -48,7 +48,7 @@ double* read_matrix(int *size, char *name){
 		return NULL;
 	}//exit if file is empty
 
-	double * matrix = new double[nrows*ncols];
+	float * matrix = new float[nrows*ncols];
 
 
 	for(int i = 0;i<input_vec.size();i++){
@@ -58,7 +58,7 @@ double* read_matrix(int *size, char *name){
 	*size = nrows;
 	return matrix;
 }
-double*** read_filter_layers(char*name, int num_filters, int filter_width, int filter_depth, double*biases){
+float*** read_filter_layers(char*name, int num_filters, int filter_width, int filter_depth, float*biases){
         int num_biases = num_filters;
         ifstream file;
         try{
@@ -69,14 +69,14 @@ double*** read_filter_layers(char*name, int num_filters, int filter_width, int f
           return NULL;
         }
 
-        double***filter_layers = new double**[num_filters];
+        float***filter_layers = new float**[num_filters];
 
         for(int i = 0;i<num_filters;i++){
-          filter_layers[i] = new double*[filter_depth];
-          for(int j = 0;j<filter_depth;j++)filter_layers[i][j] = new double [filter_width*filter_width];
+          filter_layers[i] = new float*[filter_depth];
+          for(int j = 0;j<filter_depth;j++)filter_layers[i][j] = new float [filter_width*filter_width];
         }
            int count = 0;
-        double value;
+        float value;
       //  cout<<num_filters<<" "<<filter_depth<<" "<<filter_width*filter_width<<endl;
         for(int i = 0;i<num_filters;i++){
           for(int j = 0;j<filter_depth;j++){
@@ -108,25 +108,25 @@ double*** read_filter_layers(char*name, int num_filters, int filter_width, int f
 
 
 int main(int argc, char** argv){
-  //./exe processed_data.txt conv1.txt conv2.txt fc1.txt fc2.txt
+  // To run the input->        ./exe processed_data.txt conv1.txt conv2.txt fc1.txt fc2.txt
 
   // input_matrix -> conv1 -> pooling -> con
 
-   double *input_matrix;
+   float *input_matrix;
    int input_row_size;
 
    input_matrix = read_matrix(&input_row_size,argv[1]);
 
 
-   double** input_layers = new double*[1];
+   float** input_layers = new float*[1];
    input_layers[0] = input_matrix;
 
-   double* biases = new double[20];
-   double *** filter_layers = read_filter_layers(argv[2],20,5,1,biases);
-   double *temp = filter_layers[0][0];// expected filter as  20 X 1 X 25
+   float* biases = new float[20];
+   float *** filter_layers = read_filter_layers(argv[2],20,5,1,biases);
+   float *temp = filter_layers[0][0];// expected filter as  20 X 1 X 25
   //conv1
 
-   double** output_layers = convolute_multiple(input_layers, 28,1,filter_layers,biases,5,20,0,false,false);
+   float** output_layers = convolute_multiple(input_layers, 28,1,filter_layers,biases,5,20,0,false,false);
   // output_layer   20 X 576
 
 
@@ -135,7 +135,7 @@ int main(int argc, char** argv){
      output_layers[i] = maxpool(2,24,2,0,output_layers[i]);
    }
 
-   biases = new double[50];
+   biases = new float[50];
    filter_layers = read_filter_layers(argv[3],50,5,20,biases);
    output_layers = convolute_multiple(output_layers,12,20,filter_layers,biases,5,50,0,false,false);
     // filter should be as 50 X 20 X 144
@@ -146,23 +146,23 @@ int main(int argc, char** argv){
      output_layers[i] = maxpool(2,8,2,0,output_layers[i]);
    }
 
-   biases = new double[500];
+   biases = new float[500];
    filter_layers = read_filter_layers(argv[4],500,4,50,biases);
    output_layers = convolute_multiple(output_layers,4,50,filter_layers,biases,4,500,0,false,true);
 
-   biases = new double[10];
+   biases = new float[10];
    filter_layers = read_filter_layers(argv[5],10,1,500,biases);
    output_layers = convolute_multiple(output_layers,1,500,filter_layers,biases,1,10,0,false,false);
    //output_layers 10 X 1 X 1 (for picorial representation) other wise it is just 10 X (1X1)
 
 
 
-   double* prob_vector = new double[10];
+   float* prob_vector = new float[10];
    for(int i = 0;i<10;i++)prob_vector[i] = output_layers[i][0];
    prob_vector = softmax(prob_vector,10);
 
-   map <double,int> probMap;
-   for(int i = 0;i<10;i++)probMap.insert(pair<double,int>(prob_vector[i],i));
+   map <float,int> probMap;
+   for(int i = 0;i<10;i++)probMap.insert(pair<float,int>(prob_vector[i],i));
 
    sort(prob_vector,prob_vector+10,std::greater<>());
