+++ pooling.h
@@ -1,8 +1,8 @@
 #pragma once
 using namespace std;
 
-double* maxpool(int filter_size,int matrix_size,int stride,double* A);
-double* maxpool(int filter_size,int matrix_size,int stride,int padding_size,double* A);
+float* maxpool(int filter_size,int matrix_size,int stride,float* A);
+float* maxpool(int filter_size,int matrix_size,int stride,int padding_size,float* A);
 
-double* averagepool(int filter_size,int matrix_size,int stride,double* A);
-double* averagepool(int filter_size,int matrix_size,int stride,int padding_size,double* A);
+float* averagepool(int filter_size,int matrix_size,int stride,float* A);
+float* averagepool(int filter_size,int matrix_size,int stride,int padding_size,float* A);
