+++ sigmoid.h
@@ -1,4 +1,4 @@
 #pragma once
 using namespace:: std;
 
-double* sigmoid(double* input_vector,int size);
+float* sigmoid(float* input_vector,int size);
